requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-debug-toolbar==4.0.0
django-extensions==3.2.1
django-filter==24.2
django-htmx==1.17.3
django-widget-tweaks==1.5.0
factory-boy==3.3.0
Faker==18.6.0
idna==3.4
iniconfig==2.0.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pluggy==1.5.0
pycparser==2.21
PyJWT==2.6.0
pytest==8.2.0
pytest-django==4.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
requests==2.29.0
requests-oauthlib==1.3.1
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tzdata==2023.3
urllib3==1.26.15
wheel==0.43.0
